requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asciitree==0.3.3
astroid==3.3.6
attrs==25.1.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
coverage==7.6.9
cycler==0.12.1
Deprecated==1.2.18
dill==0.3.9
distinctipy==1.3.4
fastapi==0.115.8
fasteners==0.19
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
h5py==3.12.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
isosplit6==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.26.0
neo==0.14.0
networkx==3.4.2
numba==0.61.0
numcodecs==0.15.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
probeinterface==0.2.25
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pylint==3.3.2
pyparsing==3.2.1
pyqtgraph==0.13.7
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
quantities==0.16.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.2
shiboken6==6.8.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
streamlit==1.43.1
sympy==1.13.1
tdt==0.6.7
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zarr==2.17.2
